split executors
